feat: added FileHelpers utility class + downloadFile function moved from ExportLM to FileHelpers
